# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/service/service_add_grpc": {
            "post": {
                "description": "grpc服务添加",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "grpc服务添加",
                "operationId": "/service/service_add_grpc",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceAddGrpcInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_add_http": {
            "post": {
                "description": "添加HTTP服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "添加HTTP服务",
                "operationId": "/service/service_add_http",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceAddHTTPInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_add_tcp": {
            "post": {
                "description": "tcp服务添加",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "tcp服务添加",
                "operationId": "/service/service_add_tcp",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceAddTcpInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_delete": {
            "get": {
                "description": "服务删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "服务删除",
                "operationId": "/service/service_delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_list": {
            "get": {
                "description": "服务列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "服务列表",
                "operationId": "/service/service_list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "info",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页个数",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page_no",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ServiceListOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_update_grpc": {
            "post": {
                "description": "grpc服务更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "grpc服务更新",
                "operationId": "/service/service_update_grpc",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceUpdateGrpcInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_update_http": {
            "post": {
                "description": "修改HTTP服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "修改HTTP服务",
                "operationId": "/service/service_update_http",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceUpdateHTTPInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_update_tcp": {
            "post": {
                "description": "tcp服务更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "tcp服务更新",
                "operationId": "/service/service_update_tcp",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ServiceUpdateTcpInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ServiceAddGrpcInput": {
            "type": "object",
            "required": [
                "ip_list",
                "port",
                "service_desc",
                "service_name",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "type": "string"
                },
                "clientip_flow_limit": {
                    "type": "integer"
                },
                "forbid_list": {
                    "type": "string"
                },
                "header_transfor": {
                    "type": "string"
                },
                "ip_list": {
                    "type": "string"
                },
                "open_auth": {
                    "type": "integer"
                },
                "port": {
                    "type": "integer",
                    "maximum": 8999,
                    "minimum": 8001
                },
                "round_type": {
                    "type": "integer"
                },
                "service_desc": {
                    "type": "string"
                },
                "service_flow_limit": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "weight_list": {
                    "type": "string"
                },
                "white_host_name": {
                    "type": "string"
                },
                "white_list": {
                    "type": "string"
                }
            }
        },
        "dto.ServiceAddHTTPInput": {
            "type": "object",
            "required": [
                "ip_list",
                "rule",
                "service_desc",
                "service_name",
                "url_rewrite",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "description": "黑名单ip",
                    "type": "string",
                    "example": ""
                },
                "clientip_flow_limit": {
                    "description": "\u0008客户端ip限流",
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                },
                "header_transfor": {
                    "description": "header转换",
                    "type": "string",
                    "example": ""
                },
                "ip_list": {
                    "description": "ip列表",
                    "type": "string",
                    "example": "0"
                },
                "need_https": {
                    "description": "支持https",
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0,
                    "example": 0
                },
                "need_strip_uri": {
                    "description": "启用strip_uri",
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0,
                    "example": 0
                },
                "need_websocket": {
                    "description": "是否支持websocket",
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0,
                    "example": 0
                },
                "open_auth": {
                    "description": "关键词",
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0,
                    "example": 0
                },
                "round_type": {
                    "description": "轮询方式",
                    "type": "integer",
                    "maximum": 3,
                    "minimum": 0,
                    "example": 0
                },
                "rule": {
                    "description": "域名或者前缀",
                    "type": "string",
                    "example": ""
                },
                "rule_type": {
                    "description": "接入类型",
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0,
                    "example": 0
                },
                "service_desc": {
                    "description": "服务描述",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": ""
                },
                "service_flow_limit": {
                    "description": "服务端限流",
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                },
                "service_name": {
                    "description": "服务名",
                    "type": "string",
                    "example": ""
                },
                "upstream_connect_timeout": {
                    "description": "建立连接超时, 单位s",
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                },
                "upstream_header_timeout": {
                    "description": "获取header超时, 单位s",
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                },
                "upstream_idle_timeout": {
                    "description": "链接最大空闲时间, 单位s",
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                },
                "upstream_max_idle": {
                    "description": "最大空闲链接数",
                    "type": "integer",
                    "minimum": 0,
                    "example": 10
                },
                "url_rewrite": {
                    "description": "url重写功能",
                    "type": "string",
                    "example": ""
                },
                "weight_list": {
                    "description": "\u0008权重列表",
                    "type": "string",
                    "example": ""
                },
                "white_list": {
                    "description": "白名单ip",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "dto.ServiceAddTcpInput": {
            "type": "object",
            "required": [
                "ip_list",
                "port",
                "service_desc",
                "service_name",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "type": "string"
                },
                "clientip_flow_limit": {
                    "type": "integer"
                },
                "forbid_list": {
                    "type": "string"
                },
                "header_transfor": {
                    "type": "string"
                },
                "ip_list": {
                    "type": "string"
                },
                "open_auth": {
                    "type": "integer"
                },
                "port": {
                    "type": "integer",
                    "maximum": 8999,
                    "minimum": 8001
                },
                "round_type": {
                    "type": "integer"
                },
                "service_desc": {
                    "type": "string"
                },
                "service_flow_limit": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "weight_list": {
                    "type": "string"
                },
                "white_host_name": {
                    "type": "string"
                },
                "white_list": {
                    "type": "string"
                }
            }
        },
        "dto.ServiceListItemOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "load_type": {
                    "type": "integer"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "service_addr": {
                    "type": "string"
                },
                "service_desc": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "total_node": {
                    "type": "integer"
                }
            }
        },
        "dto.ServiceListOutput": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ServiceListItemOutput"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "dto.ServiceUpdateGrpcInput": {
            "type": "object",
            "required": [
                "id",
                "ip_list",
                "port",
                "service_desc",
                "service_name",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "type": "string"
                },
                "clientip_flow_limit": {
                    "type": "integer"
                },
                "forbid_list": {
                    "type": "string"
                },
                "header_transfor": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip_list": {
                    "type": "string"
                },
                "open_auth": {
                    "type": "integer"
                },
                "port": {
                    "type": "integer",
                    "maximum": 8999,
                    "minimum": 8001
                },
                "round_type": {
                    "type": "integer"
                },
                "service_desc": {
                    "type": "string"
                },
                "service_flow_limit": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "weight_list": {
                    "type": "string"
                },
                "white_host_name": {
                    "type": "string"
                },
                "white_list": {
                    "type": "string"
                }
            }
        },
        "dto.ServiceUpdateHTTPInput": {
            "type": "object",
            "required": [
                "ip_list",
                "rule",
                "service_desc",
                "service_name",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "description": "黑名单ip",
                    "type": "string",
                    "example": ""
                },
                "clientip_flow_limit": {
                    "description": "\u0008客户端ip限流",
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                },
                "header_transfor": {
                    "description": "header转换",
                    "type": "string",
                    "example": ""
                },
                "id": {
                    "type": "integer",
                    "example": 0
                },
                "ip_list": {
                    "description": "ip列表",
                    "type": "string",
                    "example": "0"
                },
                "need_https": {
                    "description": "支持https",
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0,
                    "example": 0
                },
                "need_strip_uri": {
                    "description": "启用strip_uri",
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0,
                    "example": 0
                },
                "need_websocket": {
                    "description": "是否支持websocket",
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0,
                    "example": 0
                },
                "open_auth": {
                    "description": "关键词",
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0,
                    "example": 0
                },
                "round_type": {
                    "description": "轮询方式",
                    "type": "integer",
                    "maximum": 3,
                    "minimum": 0,
                    "example": 0
                },
                "rule": {
                    "description": "域名或者前缀",
                    "type": "string",
                    "example": ""
                },
                "rule_type": {
                    "description": "接入类型",
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0,
                    "example": 0
                },
                "service_desc": {
                    "description": "服务描述",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": ""
                },
                "service_flow_limit": {
                    "description": "服务端限流",
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                },
                "service_name": {
                    "description": "服务名",
                    "type": "string",
                    "example": ""
                },
                "upstream_connect_timeout": {
                    "description": "建立连接超时, 单位s",
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                },
                "upstream_header_timeout": {
                    "description": "获取header超时, 单位s",
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                },
                "upstream_idle_timeout": {
                    "description": "链接最大空闲时间, 单位s",
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                },
                "upstream_max_idle": {
                    "description": "最大空闲链接数",
                    "type": "integer",
                    "minimum": 0,
                    "example": 10
                },
                "url_rewrite": {
                    "description": "url重写功能",
                    "type": "string",
                    "example": ""
                },
                "weight_list": {
                    "description": "\u0008权重列表",
                    "type": "string",
                    "example": ""
                },
                "white_list": {
                    "description": "白名单ip",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "dto.ServiceUpdateTcpInput": {
            "type": "object",
            "required": [
                "id",
                "ip_list",
                "port",
                "service_desc",
                "service_name",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "type": "string"
                },
                "clientip_flow_limit": {
                    "type": "integer"
                },
                "forbid_list": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip_list": {
                    "type": "string"
                },
                "open_auth": {
                    "type": "integer"
                },
                "port": {
                    "type": "integer",
                    "maximum": 8999,
                    "minimum": 8001
                },
                "round_type": {
                    "type": "integer"
                },
                "service_desc": {
                    "type": "string"
                },
                "service_flow_limit": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "weight_list": {
                    "type": "string"
                },
                "white_host_name": {
                    "type": "string"
                },
                "white_list": {
                    "type": "string"
                }
            }
        },
        "middleware.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errmsg": {
                    "type": "string"
                },
                "errno": {
                    "type": "integer"
                },
                "stack": {},
                "trace_id": {}
            }
        }
    }
}